Quitar warnings de unreferenced local variable en catch. Otros detallines.

+++ source/windows/app.cpp
@@ -324,7 +324,7 @@ void CApp::doFrame() {
 
     Engine.render();*/
 
-    double frameTime = time_since_last_render.elapsedAndReset();
-    Engine.update((float)frameTime);
+    float frameTime = time_since_last_render.elapsedAndReset();
+    Engine.update(frameTime);
     Engine.render();
 }
